requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
backoff==2.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
eval-type-backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.8
filetype==1.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
msrest==0.7.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.61.0
packaging==24.2
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-iso639==2025.1.28
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
rapidfuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
unstructured==0.16.17
unstructured-client==0.29.0
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
wrapt==1.17.2
